Get rid of geopm_internal.h requirements from base build
- Need to be able to build the base directory with only
  installed headers.
- Add a header for field operations.
- These functions enable signals to be interpreted as bit fields.
