+++ service/src/geopm_field.h
@@ -30,57 +30,43 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "config.h"
-#include "SSTSignal.hpp"
+#ifndef GEOPM_FIELD_H_INCLUDE
+#define GEOPM_FIELD_H_INCLUDE
+#include <stdint.h>
+#include <string.h>
 
-#include "SSTIO.hpp"
-#include "geopm_field.h"
+#ifdef __cplusplus
+extern "C" {
+#endif
 
-namespace geopm
+/*!
+ * @brief Convert a signal that is implicitly a 64-bit field
+ *        especially useful for converting region IDs.
+ * @param [in] signal value returned by PlatformIO::sample() or
+ *        PlatformIO::read_signal() for a signal with a name that
+ *        ends with the '#' character.
+ */
+static inline uint64_t geopm_signal_to_field(double signal)
 {
-
-    SSTSignal::SSTSignal(std::shared_ptr<geopm::SSTIO> sstio, SignalType signal_type,
-                         int cpu_idx, uint16_t command, uint16_t subcommand,
-                         uint32_t subcommand_arg,      // write_value
-                         uint32_t interface_parameter) // mbox_interface_param
-        : m_sstio(sstio)
-        , m_signal_type(signal_type)
-        , m_cpu_idx(cpu_idx)
-        , m_command(command)
-        , m_subcommand(subcommand)
-        , m_subcommand_arg(subcommand_arg)
-        , m_batch_idx(-1)
-    {
-    }
-
-    void SSTSignal::setup_batch(void)
-    {
-        if (m_batch_idx == -1) {
-            if (m_signal_type == M_MMIO) {
-                m_batch_idx = m_sstio->add_mmio_read(m_cpu_idx, m_subcommand_arg);
-            }
-            else {
-                m_batch_idx = m_sstio->add_mbox_read(
-                    m_cpu_idx, m_command, m_subcommand, m_subcommand_arg);
-            }
-        }
-    }
-
-    double SSTSignal::sample(void)
-    {
-        return geopm_field_to_signal(m_sstio->sample(m_batch_idx));
-    }
-
-    double SSTSignal::read(void) const
-    {
-        uint32_t ret;
-        if (m_signal_type == M_MMIO) {
-            ret = m_sstio->read_mmio_once(m_cpu_idx, m_subcommand_arg);
-        }
-        else {
-            ret = m_sstio->read_mbox_once(m_cpu_idx, m_command, m_subcommand,
-                                          m_subcommand_arg);
-        }
-        return geopm_field_to_signal(ret);
-    }
+    uint64_t result;
+    memcpy(&result, &signal, sizeof(result));
+    return result;
 }
+
+/*!
+ * @brief Convert a 64-bit field into a double representation
+ *        appropriate for a signal returned by an IOGroup.
+ * @param [in] field Arbitrary 64-bit field to be stored in a
+ *        double precision value.
+ */
+static inline double geopm_field_to_signal(uint64_t field)
+{
+    double result;
+    memcpy(&result, &field, sizeof(result));
+    return result;
+}
+
+#ifdef __cplusplus
+}
+#endif
+#endif
