--- service/src/geopm_internal.h
@@ -1,214 +1,0 @@
-/*
- * Copyright (c) 2015 - 2022, Intel Corporation
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in
- *       the documentation and/or other materials provided with the
- *       distribution.
- *
- *     * Neither the name of Intel Corporation nor the names of its
- *       contributors may be used to endorse or promote products derived
- *       from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY LOG OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#ifndef GEOPM_INTERNAL_H_INCLUDE
-#define GEOPM_INTERNAL_H_INCLUDE
-#include <stdint.h>
-#include <limits.h>
-
-#include "geopm_time.h"
-#include "geopm_hint.h"
-
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/***************************/
-/* APPLICATION REGION HASH */
-/***************************/
-enum geopm_region_hash_e {
-    GEOPM_REGION_HASH_INVALID  = 0x0ULL,
-    GEOPM_REGION_HASH_UNMARKED = 0x725e8066ULL, /* Note the value is the geopm_crc32_str() of the stringified enum */
-    GEOPM_U64_SENTINEL_REGION_HASH = UINT64_MAX, /* Force enum type to uint64_t */
-};
-
-/*!
- * @brief Enum for internally defined region hashes.
- * Note the value is the geopm_crc32_str() of the stringified enum
- */
-enum geopm_region_hash_epoch_e {
-    GEOPM_REGION_HASH_EPOCH = 0x66c91423ULL,
-    GEOPM_REGION_HASH_APP = 0x9d540c53ULL,
-};
-
-/*!
- * Region id enums go from bit 63 and work their way down.
- * Hint GEOPM_REGION_HINT_* enums in go from bit 32 and work their way up.
- * There is a possibility of a conflict sometime in the future if they overlap.
- */
-enum geopm_region_id_e {
-    GEOPM_REGION_ID_EPOCH =        1ULL << 63, /* Signaling the start of an epoch, no associated Region */
-    GEOPM_REGION_ID_MPI =          1ULL << 62, /* Execution of MPI calls */
-};
-
-enum geopm_ctl_e {
-    GEOPM_CTL_NONE,
-    GEOPM_CTL_PROCESS,
-    GEOPM_CTL_PTHREAD,
-};
-
-/*!
- * @brief Structure used to hold single profiling
- * messages obtained from the application.
- */
-struct geopm_prof_message_s {
-    /*!
-     * @brief Rank identifier.
-     */
-    int rank;
-    /*!
-     * @brief 64-bit unique application region identifier.
-     */
-    uint64_t region_id;
-    /*!
-     * @brief Time stamp of when the sample was taken.
-     */
-    struct geopm_time_s timestamp;
-    /*!
-     * @brief Progress of the rank within the current region.
-     */
-    double progress;
-};
-
-/**
- * Helper that creates the DefaultProfile signleton (if not already created)
- * and catches all exceptions.
- */
-int geopm_prof_init(void);
-
-/*********************************/
-/* APPLICATION PROFILING INSIGHT */
-/*********************************/
-static inline uint64_t geopm_region_id_hash(uint64_t region_id)
-{
-    uint64_t ret = ((region_id << 32) >> 32);
-
-    if (GEOPM_REGION_HASH_INVALID == ret) {
-        ret = GEOPM_REGION_HASH_UNMARKED;
-    }
-    return ret;
-}
-
-static inline int geopm_region_id_is_epoch(uint64_t region_id)
-{
-    return (region_id & GEOPM_REGION_ID_EPOCH) ? 1 : 0;
-}
-
-static inline int geopm_region_id_is_mpi(uint64_t region_id)
-{
-    return (region_id & GEOPM_REGION_ID_MPI) ? 1 : 0;
-}
-
-static inline int geopm_region_id_is_nested(uint64_t region_id)
-{
-    return (geopm_region_id_is_mpi(region_id) && geopm_region_id_hash(region_id));
-}
-
-static inline uint64_t geopm_region_id_parent(uint64_t region_id)
-{
-    return (geopm_region_id_is_nested(region_id) ? geopm_region_id_hash(region_id) : GEOPM_REGION_HASH_UNMARKED);
-}
-
-static inline uint64_t geopm_region_id_set_mpi(uint64_t region_id)
-{
-    return (region_id | GEOPM_REGION_ID_MPI);
-}
-
-static inline uint64_t geopm_region_id_unset_mpi(uint64_t region_id)
-{
-    return (region_id & (~GEOPM_REGION_ID_MPI));
-}
-
-static inline uint64_t geopm_region_id_set_hint(uint64_t hint_type, uint64_t region_id)
-{
-    return (region_id | hint_type);
-}
-
-static inline uint64_t geopm_region_id_unset_hint(uint64_t hint_type, uint64_t region_id)
-{
-    return (region_id & (~hint_type));
-}
-
-static inline int geopm_region_id_hint_is_equal(uint64_t hint_type, uint64_t region_id)
-{
-    return (region_id & hint_type) ? 1 : 0;
-}
-
-static inline uint64_t geopm_region_id_hint(uint64_t region_id)
-{
-    uint64_t ret;
-    if (GEOPM_REGION_HASH_UNMARKED == region_id) {
-        ret = GEOPM_REGION_HINT_UNKNOWN;
-    }
-    else if (geopm_region_id_is_mpi(region_id)) {
-        ret = GEOPM_REGION_HINT_NETWORK;
-    }
-    else {
-        ret = region_id & GEOPM_MASK_REGION_HINT;
-        if (!ret) {
-            ret = GEOPM_REGION_HINT_UNKNOWN;
-        }
-    }
-    return ret;
-}
-
-/*!
- * @brief Convert a signal that is implicitly a 64-bit field
- *        especially useful for converting region IDs.
- * @param [in] signal value returned by PlatformIO::sample() or
- *        PlatformIO::read_signal() for a signal with a name that
- *        ends with the '#' character.
- */
-static inline uint64_t geopm_signal_to_field(double signal)
-{
-    uint64_t result;
-    memcpy(&result, &signal, sizeof(result));
-    return result;
-}
-
-/*!
- * @brief Convert a 64-bit field into a double representation
- *        appropriate for a signal returned by an IOGroup.
- * @param [in] field Arbitrary 64-bit field to be stored in a
- *        double precision value.
- */
-static inline double geopm_field_to_signal(uint64_t field)
-{
-    double result;
-    memcpy(&result, &field, sizeof(result));
-    return result;
-}
-
-#ifdef __cplusplus
-}
-#endif
-#endif
