+++ service/test/PlatformIOTest.cpp
@@ -39,7 +39,6 @@
 #include "gtest/gtest.h"
 #include "gmock/gmock.h"
 
-#include "geopm_internal.h"
 #include "geopm_hash.h"
 #include "PlatformIOImp.hpp"
 #include "geopm/IOGroup.hpp"
