+++ service/test/SSTIOGroupTest.cpp
@@ -37,7 +37,6 @@
 #include "gtest/gtest.h"
 #include "gmock/gmock.h"
 
-#include "geopm_internal.h"
 #include "geopm/Helper.hpp"
 #include "SSTIOGroup.hpp"
 #include "MockPlatformTopo.hpp"
