+++ src/ApplicationSampler.cpp
@@ -53,7 +53,7 @@
 #include "record.hpp"
 #include "geopm_debug.hpp"
 #include "geopm_hash.h"
-#include "geopm_internal.h"
+#include "geopm_hint.h"
 
 namespace geopm
 {
