+++ src/ApplicationStatus.cpp
@@ -35,11 +35,12 @@
 
 #include <cmath>
 
-#include "geopm_internal.h"
 #include "geopm/SharedMemory.hpp"
 #include "geopm/Helper.hpp"
 #include "geopm/Exception.hpp"
 #include "geopm_debug.hpp"
+#include "geopm_hash.h"
+#include "geopm_hint.h"
 
 
 namespace geopm
