+++ src/FrequencyMapAgent.cpp
@@ -41,7 +41,6 @@
 
 #include "geopm/json11.hpp"
 
-#include "geopm_internal.h"
 #include "geopm_hash.h"
 
 #include "Environment.hpp"
