+++ src/OMPT.cpp
@@ -41,7 +41,6 @@
 #include <algorithm>
 
 #include "geopm_prof.h"
-#include "geopm_internal.h"
 #include "geopm_sched.h"
 #include "geopm_error.h"
 #include "Profile.hpp"
@@ -49,6 +48,8 @@
 #include "Environment.hpp"
 #include "geopm/Exception.hpp"
 #include "OMPT.hpp"
+#include "geopm_hash.h"
+#include "geopm_hint.h"
 
 #include "config.h"
