+++ test/CSVTest.cpp
@@ -43,7 +43,7 @@
 #include "CSV.hpp"
 #include "geopm_version.h"
 #include "geopm_hash.h"
-#include "geopm_internal.h"
+#include "geopm_field.h"
 #include "config.h"
