+++ test/EnvironmentTest.cpp
@@ -39,7 +39,8 @@
 
 #include "gtest/gtest.h"
 #include "geopm_error.h"
-#include "geopm_internal.h"
+#include "geopm_field.h"
+#include "geopm_prof.h"
 #include "Environment.hpp"
 #include "geopm/Exception.hpp"
 #include "geopm/Helper.hpp"
