+++ test/SampleAggregatorTest.cpp
@@ -44,8 +44,7 @@
 #include "geopm/IOGroup.hpp"
 #include "geopm_prof.h"
 #include "geopm_hash.h"
-#include "geopm_internal.h"
-#include "geopm_hash.h"
+
 #include "geopm_test.hpp"
 
 using geopm::SampleAggregator;
